requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
langchain_nvidia_ai_endpoints
langchain_chroma
langchain_huggingface
pypdf
pytesseract
Pillow
gTTS
speechrecognition
llama-index
llama-index-llms-anthropic
llama-index-llms-deepseek
llama-index-llms-gemini
llama-index-llms-mistralai
llama-index-llms-openai
